Fixes OZ Audit 2 N03 and N04

+++ contracts/external/yearn/yVaultInterface.sol
@@ -5,11 +5,11 @@ import "@openzeppelin/contracts-ethereum-package/contracts/token/ERC20/IERC20.so
 interface yVaultInterface is IERC20 {
     function token() external view returns (IERC20);
 
-    function balance() external view returns (uint);
+    function balance() external view returns (uint256);
     
-    function deposit(uint _amount) external;
+    function deposit(uint256 _amount) external;
     
-    function withdraw(uint _shares) external;
+    function withdraw(uint256 _shares) external;
     
-    function getPricePerFullShare() external view returns (uint);
+    function getPricePerFullShare() external view returns (uint256);
 }
